# frontend/docs-backend/openapi_QUYNH_v8.0.0.yaml
...
openapi: 3.1.0
info:
  title: Quynh Web API
  description: Cổng thông tin điện tử nội bộ Lữ đoàn Thông tin 21. Xem lịch sử thay
    đổi ở openapi.CHANGELOG.md.
  version: 8.0.0
paths:
...
  /announcements:
    post:
      tags:
      - announcements
      summary: Create Announcement
      operationId: create_announcement_announcements_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnnouncementCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnnouncementOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - announcements
      summary: List Announcements
      operationId: list_announcements_announcements_get
      security:
      - HTTPBearer: []
      parameters:
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 100
          title: Limit
      - name: priority
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Priority
      - name: pinned_only
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Pinned Only
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AnnouncementOut'
                title: Response List Announcements Announcements Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /announcements/{ann_id}:
    get:
      tags:
      - announcements
      summary: Get Announcement
      operationId: get_announcement_announcements__ann_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: ann_id
        in: path
        required: true
        schema:
          type: integer
          title: Ann Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnnouncementOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - announcements
      summary: Update Announcement
      operationId: update_announcement_announcements__ann_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: ann_id
        in: path
        required: true
        schema:
          type: integer
          title: Ann Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnnouncementCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnnouncementOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - announcements
      summary: Delete Announcement
      operationId: delete_announcement_announcements__ann_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: ann_id
        in: path
        required: true
        schema:
          type: integer
          title: Ann Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AnnouncementCreate:
      properties:
        title:
          type: string
          maxLength: 255
          title: Title
        content:
          type: string
          title: Content
        priority:
          type: string
          enum:
          - thap
          - binh_thuong
          - cao
          - khan
          title: Priority
          default: binh_thuong
        is_pinned:
          type: boolean
          title: Is Pinned
          default: false
        is_public:
          type: boolean
          title: Is Public
          default: false
        starts_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Starts At
        ends_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Ends At
      type: object
      required:
      - title
      - content
      title: AnnouncementCreate
    AnnouncementOut:
      properties:
        id:
          type: integer
          title: Id
        title:
          type: string
          title: Title
        content:
          type: string
          title: Content
        priority:
          type: string
          enum:
          - thap
          - binh_thuong
          - cao
          - khan
          title: Priority
        is_pinned:
          type: boolean
          title: Is Pinned
        is_public:
          type: boolean
          title: Is Public
        starts_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Starts At
        ends_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Ends At
        author_id:
          type: integer
          title: Author Id
        author_full_name:
          type: string
          title: Author Full Name
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - title
      - content
      - priority
      - is_pinned
      - is_public
      - starts_at
      - ends_at
      - author_id
      - author_full_name
      - created_at
      title: AnnouncementOut
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer